+++ tests/store/crud/test_create.py
@@ -1,7 +1,8 @@
 from typing import List
 
 from tests.mocks.store_mock import MockStore
-from trailblazer.store.models import Analysis
+from trailblazer.store.filters.user_filters import apply_user_filter, UserFilter
+from trailblazer.store.models import Analysis, User
 
 
 def test_add_pending_analysis(raw_analyses: List[dict], store: MockStore, user_email: str):
@@ -30,3 +31,21 @@ def test_add_pending_analysis(raw_analyses: List[dict], store: MockStore, user_e
     )
     assert new_analysis and stored_analysis
     assert stored_analysis == new_analysis
+
+
+def test_add_user(store: MockStore, user_email: str, username: str):
+    """Test adding a user to the database."""
+    # GIVEN an empty database
+    assert not store.get_query(table=User).first()
+
+    # WHEN adding a new user
+    new_user: User = store.add_user(name=username, email=user_email)
+
+    # THEN it should be stored in the database
+    user: User = apply_user_filter(
+        filter_functions=[UserFilter.FILTER_BY_EMAIL],
+        users=store.get_query(table=User),
+        email=user_email,
+    ).first()
+    assert new_user
+    assert user == new_user
