+++ tests/store/crud/test_update.py
@@ -6,24 +6,6 @@ from trailblazer.store.filters.user_filters import apply_user_filter, UserFilter
 from trailblazer.store.models import User, Analysis
 
 
-def test_add_user(store: MockStore, user_email: str, username: str):
-    """Test adding a user to the database."""
-    # GIVEN an empty database
-    assert not store.get_query(table=User).first()
-
-    # WHEN adding a new user
-    new_user: User = store.add_user(name=username, email=user_email)
-
-    # THEN it should be stored in the database
-    user: User = apply_user_filter(
-        filter_functions=[UserFilter.FILTER_BY_EMAIL],
-        users=store.get_query(table=User),
-        email=user_email,
-    ).first()
-    assert new_user
-    assert user == new_user
-
-
 def test_update_user_is_archived(user_store: MockStore, user_email: str):
     """Test updating user is archived attribute."""
     # GIVE a database and a not archived user
